app/requirements.txt
flask
#boto3
python-dotenv
uwsgi
ruamel.yaml
